pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "cpr-josephson-analysis"
version = "1.0.0"
description = "High-performance Josephson junction current-phase relation analysis suite"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "CPR Team", email = "support@cpr.dev"}
]
maintainers = [
    {name = "CPR Team", email = "support@cpr.dev"}
]
keywords = ["josephson junction", "superconductivity", "physics", "data analysis", "numba", "parallel processing"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.5.0",
    "pandas>=1.3.0",
    "astropy>=5.0.0",
    "numba>=0.56.0",
    "psutil>=5.8.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "pre-commit>=2.15.0",
    "mypy>=0.910",
    "pytest-cov>=2.12.0",
]
performance = [
    "fireducks-pandas",
]
all = ["cpr-josephson-analysis[dev,performance]"]

[project.scripts]
cpr-process = "cpr.main_processor:main"

[project.urls]
Homepage = "https://github.com/aa8919aa3/CPR"
Documentation = "https://github.com/aa8919aa3/CPR/wiki"
Repository = "https://github.com/aa8919aa3/CPR"
"Bug Tracker" = "https://github.com/aa8919aa3/CPR/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "numpy.*",
    "scipy.*",
    "matplotlib.*",
    "pandas.*",
    "astropy.*",
    "numba.*",
    "psutil.*",
    "seaborn.*",
    "plotly.*",
]
ignore_missing_imports = true
